# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: cncf-harbor-maintainers@lists.cncf.io
  description: |
    ## Overview

    This API must be implemented in order to register a new artifact scanner in [Harbor](https://goharbor.io) registry.

    The [/scan](#operation/AcceptScanRequest) and [/scan/{scan_request_id}/report](#operation/GetScanReport) operations are responsible for the actual scanning and return a scan report that is visible in the Harbor web console.

    The [/scan](#operation/AcceptScanRequest) operation is asynchronous. It should enqueue the job for processing a scan request and return the identifier. This allows Harbor to poll a corresponding scan report with the
    [/scan/{scan_request_id}/report](#operation/GetScanReport) operation. Harbor will call the
    [/scan/{scan_request_id}/report](#operation/GetScanReport) operation periodically periodically until it returns 200 or 500 status codes.

    The [/metadata](#operation/GetMetadata) operation allows a Harbor admin to configure and register a scanner
    and discover its capabilities.

    ## Supported consumed MIME types

    - `application/vnd.oci.image.manifest.v1+json`
    - `application/vnd.docker.distribution.manifest.v2+json`

    ## Supported produced MIME types

    - `application/vnd.scanner.adapter.vuln.report.harbor+json; version=1.0`
    - `application/vnd.scanner.adapter.vuln.report.raw`
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Harbor Scanner Adapter API
  version: "1.0"
servers:
- url: /api/v1
security:
- BasicAuth: []
- BearerAuth: []
paths:
  /metadata:
    get:
      description: |
        Used to fetch scanner's metadata and capabilities. The operation is invoked to build an index of scanners
        capable of analysing a given type of artifacts and making sure that scan reports can be parsed and rendered.
      operationId: GetMetadata
      responses:
        "200":
          content:
            application/vnd.scanner.adapter.metadata+json; version=1.0:
              schema:
                $ref: '#/components/schemas/ScannerAdapterMetadata'
          description: Scanner's metadata and capabilities
        "500":
          content:
            application/vnd.scanner.adapter.error+json; version=1.0:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      summary: Get scanner metadata
      tags:
      - Scanner
  /scan:
    post:
      description: |
        A non-blocking operation which enqueues a scan job and returns immediately. It returns a unique
        identifier which can be used to poll for generated scan reports by Harbor.
      operationId: AcceptScanRequest
      requestBody:
        content:
          application/vnd.scanner.adapter.scan.request+json; version=1.0:
            schema:
              $ref: '#/components/schemas/ScanRequest'
        description: |
          Contains data required to pull the given artifact and save it for scanning in the file system or any other
          location accessible to the scanner.
      responses:
        "202":
          content:
            application/vnd.scanner.adapter.scan.response+json; version=1.0:
              schema:
                $ref: '#/components/schemas/ScanResponse'
          description: Scan request accepted
        "400":
          content:
            application/vnd.scanner.adapter.error+json; version=1.0:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Received invalid JSON or the wrong type of JSON values
        "422":
          content:
            application/vnd.scanner.adapter.error+json; version=1.0:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Received invalid field
        "500":
          content:
            application/vnd.scanner.adapter.error+json; version=1.0:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      summary: Accept artifact scanning request
      tags:
      - Scanner
  /scan/{scan_request_id}/report:
    get:
      description: |
        Get a scan report for the given scan request identifier.

        Clients will periodically poll this operation and check `$response.status` until its value equals `200` or `500`.
      operationId: GetScanReport
      parameters:
      - description: The identifier of the corresponding scan request
        explode: false
        in: path
        name: scan_request_id
        required: true
        schema:
          $ref: '#/components/schemas/ScanRequestId'
        style: simple
      - example: application/vnd.scanner.adapter.vuln.report.harbor+json; version=1.0
        explode: false
        in: header
        name: Accept
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/vnd.scanner.adapter.vuln.report.harbor+json; version=1.0:
              schema:
                $ref: '#/components/schemas/HarborVulnerabilityReport'
            application/vnd.scanner.adapter.vuln.report.raw:
              schema:
                example: |
                  {
                    "vendor_specific": "vulnerabilities_report"
                  }
                type: string
          description: Scan report
        "302":
          description: Status indicating the scan report is being generated and the
            request should be retried.
          headers:
            Refresh-After:
              description: Indicates the interval after which the request should be
                retried.
              explode: false
              schema:
                type: integer
              style: simple
        "404":
          description: Cannot find the corresponding scan request identifier
        "500":
          content:
            application/vnd.scanner.adapter.error+json; version=1.0:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      summary: Get scan report
      tags:
      - Scanner
components:
  schemas:
    Scanner:
      description: |
        Basic scanner properties such as name, vendor, and version.
      properties:
        name:
          description: The name of the scanner.
          example: Trivy
          type: string
        vendor:
          description: The name of the scanner's provider.
          example: Aqua Security
          type: string
        version:
          description: The version of the scanner.
          example: 0.4.0
          type: string
      type: object
    ScannerAdapterMetadata:
      description: |
        Represents metadata of a Scanner Adapter which allows Harbor to lookup a scanner capable
        of scanning a given Artifact stored in its registry and making sure that it
        can interpret a returned result.
      properties:
        scanner:
          $ref: '#/components/schemas/Scanner'
        capabilities:
          items:
            $ref: '#/components/schemas/ScannerCapability'
          type: array
        properties:
          additionalProperties:
            type: string
          description: |
            A set of custom properties that can further describe capabilities of a given scanner.
          example:
            harbor.scanner-adapter/scanner-type: os-package-vulnerability
            harbor.scanner-adapter/vulnerability-database-updated-at: 2019-08-13T08:16:33.345Z
          type: object
      required:
      - capabilities
      - scanner
      type: object
    ScannerProperties:
      additionalProperties:
        type: string
      description: |
        A set of custom properties that can further describe capabilities of a given scanner.
      example:
        harbor.scanner-adapter/scanner-type: os-package-vulnerability
        harbor.scanner-adapter/vulnerability-database-updated-at: 2019-08-13T08:16:33.345Z
      type: object
    ScannerCapability:
      description: |
        Capability consists of the set of recognized artifact MIME types and the set of scanner report MIME types.
        For example, a scanner capable of analyzing Docker images and producing a vulnerabilities report recognizable
        by Harbor web console might be represented with the following capability:
        - consumes MIME types:
          - `application/vnd.oci.image.manifest.v1+json`
          - `application/vnd.docker.distribution.manifest.v2+json`
        - produces MIME types:
          - `application/vnd.scanner.adapter.vuln.report.harbor+json; version=1.0`
      properties:
        consumes_mime_types:
          description: |
            The set of MIME types of the artifacts supported by the scanner to produce the reports specified in the "produces_mime_types". A given
            mime type should only be present in one capability item.
          example:
          - application/vnd.oci.image.manifest.v1+json
          - application/vnd.docker.distribution.manifest.v2+json
          items:
            type: string
          type: array
        produces_mime_types:
          description: |
            The set of MIME types of reports generated by the scanner for the consumes_mime_types of the same capability record.
          example:
          - application/vnd.scanner.adapter.vuln.report.harbor+json; version=1.0
          items:
            type: string
          type: array
      required:
      - consumes_mime_types
      - produces_mime_types
      type: object
    ScanRequest:
      properties:
        registry:
          $ref: '#/components/schemas/Registry'
        artifact:
          $ref: '#/components/schemas/Artifact'
      required:
      - artifact
      - registry
      type: object
    ScanResponse:
      properties:
        id:
          description: |
            A unique identifier returned by the [/scan](#/operation/AcceptScanRequest] operations. The format of the
            identifier is not imposed but it should be unique enough to prevent collisons when polling for scan reports.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          type: string
      required:
      - id
    ScanRequestId:
      description: |
        A unique identifier returned by the [/scan](#/operation/AcceptScanRequest] operations. The format of the
        identifier is not imposed but it should be unique enough to prevent collisons when polling for scan reports.
      example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      type: string
    Registry:
      properties:
        url:
          description: A base URL or the Docker Registry v2 API.
          example: https://core.harbor.domain
          format: url
          type: string
        authorization:
          description: |
            An optional value of the HTTP Authorization header sent with each request to the Docker Registry v2 API.
            It's used to exchange Base64 encoded robot account credentials to a short lived JWT access token which
            allows the underlying scanner to pull the artifact from the Docker Registry.
          example: Basic BASE64_ENCODED_CREDENTIALS
          type: string
      type: object
    Artifact:
      properties:
        repository:
          description: The name of the Docker Registry repository containing the artifact.
          example: library/mongo
          type: string
        digest:
          description: The artifact's digest, consisting of an algorithm and hex portion.
          example: sha256:6c3c624b58dbbcd3c0dd82b4c53f04194d1247c6eebdaab7c610cf7d66709b3b
          type: string
        tag:
          description: The artifact's tag
          example: 3.14-xenial
          type: string
        mime_type:
          description: The MIME type of the artifact.
          example: application/vnd.docker.distribution.manifest.v2+json
          type: string
      type: object
    HarborVulnerabilityReport:
      properties:
        generated_at:
          format: date-time
          type: string
        artifact:
          $ref: '#/components/schemas/Artifact'
        scanner:
          $ref: '#/components/schemas/Scanner'
        severity:
          $ref: '#/components/schemas/Severity'
        vulnerabilities:
          items:
            $ref: '#/components/schemas/VulnerabilityItem'
          type: array
      type: object
    VulnerabilityItem:
      properties:
        id:
          description: The unique identifier of the vulnerability.
          example: CVE-2017-8283
          type: string
        package:
          description: |
            An operating system package containing the vulnerability.
          example: dpkg
          type: string
        version:
          description: |
            The version of the package containing the vulnerability.
          example: 1.17.27
          type: string
        fix_version:
          description: |
            The version of the package containing the fix if available.
          example: 1.18.0
          type: string
        severity:
          $ref: '#/components/schemas/Severity'
        description:
          description: |
            The detailed description of the vulnerability.
          example: |
            dpkg-source in dpkg 1.3.0 through 1.18.23 is able to use a non-GNU patch program
            and does not offer a protection mechanism for blank-indented diff hunks, which
            allows remote attackers to conduct directory traversal attacks via a crafted
            Debian source package, as demonstrated by using of dpkg-source on NetBSD.
          type: string
        links:
          description: |
            The list of links to the upstream databases with the full description of the vulnerability.
          example:
          - https://security-tracker.debian.org/tracker/CVE-2017-8283
          format: uri
          items:
            format: uri
            type: string
          type: array
      type: object
    Severity:
      description: |
        A standard scale for measuring the severity of a vulnerability.

        * `Unknown` - either a security problem that has not been assigned to a priority yet or a priority that the
          scanner did not recognize.
        * `Negligible` - technically a security problem, but is only theoretical in nature, requires a very special
          situation, has almost no install base, or does no real damage.
        * `Low` - a security problem, but is hard to exploit due to environment, requires a user-assisted attack,
          a small install base, or does very little damage.
        * `Medium` - a real security problem, and is exploitable for many people. Includes network daemon denial of
          service attacks, cross-site scripting, and gaining user privileges.
        * `High` - a real problem, exploitable for many people in a default installation. Includes serious remote denial
          of service, local root privilege escalations, or data loss.
        * `Critical` - a world-burning problem, exploitable for nearly all people in a default installation. Includes
          remote root privilege escalations, or massive data loss.
      enum:
      - Unknown
      - Negligible
      - Low
      - Medium
      - High
      - Critical
      example: Low
      type: string
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/Error'
      type: object
    Error:
      properties:
        message:
          example: Some unexpected error
          type: string
      type: object
  securitySchemes:
    BasicAuth:
      scheme: basic
      type: http
    BearerAuth:
      scheme: bearer
      type: http