# internal/model/harbor/metadata.go
/*
 * Copyright 2021 VMware, Inc.
 * SPDX-License-Identifier: Apache-2.0
 */
 
package harbor

type ScannerAdapterMetadata struct {
	Scanner      Scanner             `json:"scanner"`
	Capabilities []ScannerCapability `json:"capabilities"`
	// A set of custom properties that can further describe capabilities of a given scanner.
	Properties map[string]string `json:"properties,omitempty"`
}

// ScannerCapability - Capability consists of the set of recognized artifact MIME types and
// the set of scanner report MIME types. For example, a scanner capable of analyzing Docker images
// and producing a vulnerabilities report recognizable by Harbor web console might be represented
// with the following capability:
// - consumes MIME types:
// - `application/vnd.oci.image.manifest.v1+json`
// - `application/vnd.docker.distribution.manifest.v2+json`
// - produces MIME types:
// - `application/vnd.scanner.adapter.vuln.report.harbor+json; version=1.0`
type ScannerCapability struct {
	// The set of MIME types of the artifacts supported by the scanner to produce the reports
	// specified in the \"produces_mime_types\".
	// A given mime type should only be present in one capability item.
	ConsumesMimeTypes []string `json:"consumes_mime_types"`
	// The set of MIME types of reports generated by the scanner for the
	// consumes_mime_types of the same capability record.
	ProducesMimeTypes []string `json:"produces_mime_types"`
}

// Scanner - Basic scanner properties such as name, vendor, and version.
type Scanner struct {
	// The name of the scanner.
	Name string `json:"name,omitempty"`
	// The name of the scanner's provider.
	Vendor string `json:"vendor,omitempty"`
	// The version of the scanner.
	Version string `json:"version,omitempty"`
}